Internal change
PiperOrigin-RevId: 281063737
